Alteracoes na estrutura do namespace de Redes Neuronais a fim de refatorar o design das classes

+++ ECG.Framework/IFuncaoAtivacao.cs
@@ -7,6 +7,6 @@ namespace ECG.Framework
 {
     public interface IFuncaoAtivacao
     {
-        float Calcular(float[] entradas, float[] pesos);
+        double Calcular(double[] entradas, double[] pesos);
     }
 }
